ska-tmc/simulators/requirements.txt
numpy == 1.17.2
pytango >= 9.3.3
jsonschema >= 3.2.0
ska-ser-logging
marshmallow
fire
ska-ser-log-transactions
pytz
ska-tango-base == 0.7.2